Update accreditation and location mappings
